Finish 16-bit control fields
This updates the 16-bit vmcs control fields
to use std::enable_if and bitmanip.h.
