+++ bfvmm/include/vmcs/vmcs_intel_x64_16bit_control_fields.h
@@ -1,0 +1,123 @@
+//
+// Bareflank Hypervisor
+//
+// Copyright (C) 2015 Assured Information Security, Inc.
+// Author: Rian Quinn        <quinnr@ainfosec.com>
+// Author: Brendan Kerrigan  <kerriganb@ainfosec.com>
+// Author: Connor Davis      <davisc@ainfosec.com>
+//
+// This library is free software; you can redistribute it and/or
+// modify it under the terms of the GNU Lesser General Public
+// License as published by the Free Software Foundation; either
+// version 2.1 of the License, or (at your option) any later version.
+//
+// This library is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+// Lesser General Public License for more details.
+//
+// You should have received a copy of the GNU Lesser General Public
+// License along with this library; if not, write to the Free Software
+// Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
+
+#ifndef VMCS_INTEL_X64_16BIT_CONTROL_FIELDS_H
+#define VMCS_INTEL_X64_16BIT_CONTROL_FIELDS_H
+
+#include <vmcs/vmcs_intel_x64_state.h>
+#include <intrinsics/msrs_intel_x64.h>
+
+
+/// Intel x86_64 VMCS 16-Bit Control Fields
+///
+/// This provides the namespace for 16-bit control fields of the VMCS.
+/// See Appendix B.1.1, Vol. 3 of the Intel Software Developer's Manual
+/// for details.
+///
+
+// *INDENT-OFF*
+namespace intel_x64
+{
+namespace vmcs
+{
+
+namespace virtual_processor_identifier
+{
+    constexpr const auto addr = 0x0000000000000000UL;
+    constexpr const auto name = "virtual_processor_identifier";
+
+    inline bool exists() noexcept
+    {
+        return msrs::ia32_vmx_true_procbased_ctls::activate_secondary_controls::is_allowed1() &&
+               msrs::ia32_vmx_procbased_ctls2::enable_vpid::is_allowed1();
+    }
+
+    inline auto get()
+    { return get_vmcs_field(addr, name, exists()); }
+
+    inline auto get_if_exists(bool verbose = false) noexcept
+    { return get_vmcs_field_if_exists(addr, name, verbose, exists()); }
+
+    template<class T, class = typename std::enable_if<std::is_integral<T>::value>::type>
+    void set(T val)
+    { set_vmcs_field(val, addr, name, exists()); }
+
+    template<class T, class = typename std::enable_if<std::is_integral<T>::value>::type>
+    void set_if_exists(T val, bool verbose = false) noexcept
+    { set_vmcs_field_if_exists(val, addr, name, verbose, exists()); }
+}
+
+namespace posted_interrupt_notification_vector
+{
+    constexpr const auto addr = 0x0000000000000002UL;
+    constexpr const auto name = "posted_interrupt_notification_vector";
+
+    inline bool exists() noexcept
+    { return msrs::ia32_vmx_true_pinbased_ctls::process_posted_interrupts::is_allowed1(); }
+
+    inline auto get()
+    { return get_vmcs_field(addr, name, exists()); }
+
+    inline auto get_if_exists(bool verbose = false) noexcept
+    { return get_vmcs_field_if_exists(addr, name, verbose, exists()); }
+
+    template<class T, class = typename std::enable_if<std::is_integral<T>::value>::type>
+    void set(T val)
+    { set_vmcs_field(val, addr, name, exists()); }
+
+    template<class T, class = typename std::enable_if<std::is_integral<T>::value>::type>
+    void set_if_exists(T val, bool verbose = false) noexcept
+    { set_vmcs_field_if_exists(val, addr, name, verbose, exists()); }
+}
+
+namespace eptp_index
+{
+    constexpr const auto addr = 0x0000000000000004UL;
+    constexpr const auto name = "eptp_index";
+
+    inline bool exists() noexcept
+    {
+        return msrs::ia32_vmx_true_procbased_ctls::activate_secondary_controls::is_allowed1() &&
+               msrs::ia32_vmx_procbased_ctls2::ept_violation_ve::is_allowed1();
+    }
+
+    inline auto get()
+    { return get_vmcs_field(addr, name, exists()); }
+
+    inline auto get_if_exists(bool verbose = false) noexcept
+    { return get_vmcs_field_if_exists(addr, name, verbose, exists()); }
+
+    template<class T, class = typename std::enable_if<std::is_integral<T>::value>::type>
+    void set(T val)
+    { set_vmcs_field(val, addr, name, exists()); }
+
+    template<class T, class = typename std::enable_if<std::is_integral<T>::value>::type>
+    void set_if_exists(T val, bool verbose = false) noexcept
+    { set_vmcs_field_if_exists(val, addr, name, verbose, exists()); }
+}
+
+}
+}
+
+// *INDENT-ON*
+
+#endif
