+++ bfvmm/src/vmcs/src/vmcs_intel_x64_check_controls.cpp
@@ -23,6 +23,7 @@
 #include <view_as_pointer.h>
 #include <vmcs/vmcs_intel_x64.h>
 #include <memory_manager/memory_manager_x64.h>
+#include <vmcs/vmcs_intel_x64_16bit_control_fields.h>
 
 using namespace intel_x64;
 using namespace vmcs;
