+++ bfvmm/src/vmcs/test/test_vmcs_intel_x64.cpp
@@ -26,6 +26,7 @@
 #include <intrinsics/tss_x64.h>
 #include <vmcs/vmcs_intel_x64_resume.h>
 #include <vmcs/vmcs_intel_x64_promote.h>
+#include <vmcs/vmcs_intel_x64_16bit_control_fields.h>
 
 #include <intrinsics/rflags_x64.h>
 #include <intrinsics/crs_intel_x64.h>
