+++ bfvmm/src/vmcs/test/test_vmcs_intel_x64_check_controls.cpp
@@ -21,6 +21,7 @@
 
 #include <test.h>
 #include <memory_manager/memory_manager_x64.h>
+#include <vmcs/vmcs_intel_x64_16bit_control_fields.h>
 
 using namespace intel_x64;
